+++ Modules/Microsoft365DSC/DSCResources/MSFT_AADGroupsNamingPolicy/MSFT_AADGroupsNamingPolicy.psm1
@@ -181,6 +181,7 @@ function Set-TargetResource
     $needToUpdate = $false
     if ($Ensure -eq "Present" -and $currentPolicy.Ensure -eq "Absent")
     {
+        Write-Verbose -Message "Creating new Groups Naming Policy"
         $Policy = New-MgDirectorySetting -TemplateId '62375ab9-6b52-47ed-826b-58e47e0e304b' | Out-Null
         $needToUpdate = $true
     }
@@ -209,11 +210,12 @@ function Set-TargetResource
             }
             $index++;
         }
-
+        Write-Verbose -Message "Updating Groups Naming Policy to {$($Policy.Values -join ',')}"
         Update-MgDirectorySetting -DirectorySettingId $Policy.id -Values $Policy.Values | Out-Null
     }
     elseif ($Ensure -eq 'Absent' -and $currentPolicy.Ensure -eq 'Present')
     {
+        Write-Verbose -Message "Removing existing Groups Naming Policy {$($policy.Id)}"
         $Policy = Get-MgDirectorySetting | Where-Object -FilterScript { $_.DisplayName -eq "Group.Unified" }
         Remove-MgDirectorySetting -DirectorySettingId $policy.Id | Out-Null
     }