+++ Modules/Microsoft365DSC/DSCResources/MSFT_EXOAntiPhishPolicy/MSFT_EXOAntiPhishPolicy.psm1
@@ -455,26 +455,31 @@ function Set-TargetResource
     $ConnectionMode = New-M365DSCConnection -Workload 'ExchangeOnline' `
         -InboundParameters $PSBoundParameters
 
-    $AntiPhishPolicies = Get-AntiPhishPolicy
+    $currentInstance = Get-TargetResource @PSBoundParameters
 
-    $AntiPhishPolicy = $AntiPhishPolicies | Where-Object -FilterScript { $_.Identity -eq $Identity }
+    $PSBoundParameters.Remove("ApplicationId") | Out-Null
+    $PSBoundParameters.Remove("TenantId") | Out-Null
+    $PSBoundParameters.Remove("CertificateThumbprint") | Out-Null
+    $PSBoundParameters.Remove("CertificatePassword") | Out-Null
+    $PSBoundParameters.Remove("CertificatePath") | Out-Null
+    $PSBoundParameters.Remove("Credential") | Out-Null
 
-    if (('Present' -eq $Ensure ) -and (-not $AntiPhishPolicy))
+    if (('Present' -eq $Ensure ) -and $currentInstance.Ensure -eq 'Absent')
     {
-        New-EXOAntiPhishPolicy -AntiPhishPolicyParams $PSBoundParameters
+        Write-Verbose -Message "Creating new instance of AntiPhish Policy {$Identity}"
+        $CreateParams = $PSBoundParameters
+        $CreateParams.Remove("Ensure") | Out-Null
+        $createParams.Add("Name", $Identity)
+        $createParams.Remove("Identity") | Out-Null
+        New-AntiPhishPolicy -AntiPhishPolicyParams $PSBoundParameters
     }
-
-    if (('Present' -eq $Ensure ) -and ($AntiPhishPolicy))
+    elseif (('Present' -eq $Ensure ) -and $currentInstance.Ensure -eq 'Present')
     {
-        # The Set-AntiphishPolicy cmdlet doesn't account for more than 80% of the parameters
-        # defined by the New-AntiphishPolicy one. Therefore we need to delete the existing
-        # policy and recreate it in order to make sure the parameters all match the desired
-        # state specified;
-        Remove-AntiPhishPolicy -Identity $Identity -Confirm:$false -Force
-        New-EXOAntiPhishPolicy -AntiPhishPolicyParams $PSBoundParameters
+        Write-Verbose -MEssage "Updating existing AntiPhishPolicy {$Identity}"
+        $UpdateParams = $PSBoundParameters
+        $UpdateParams.Remove("Ensure") | Out-Null
     }
-
-    if (('Absent' -eq $Ensure ) -and ($AntiPhishPolicy))
+    elseif (('Absent' -eq $Ensure ) -and $currentInstance.Ensure -eq 'Present')
     {
         Write-Verbose -Message "Removing AntiPhishPolicy $($Identity)"
         Remove-AntiPhishPolicy -Identity $Identity -Confirm:$false -Force